requirements.txt
absl-py==0.15.0
addict==2.4.0
anyio==4.1.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==1.0.3
attrs==25.1.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
distinctipy==1.1.5
easydict==1.9
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.0.3
flatbuffers==1.12
fonttools==4.56.0
fqdn==1.5.1
fsspec==2025.3.0
gast==0.3.3
google-auth==2.38.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h11==0.14.0
h5py==2.10.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaggle==1.6.17
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib~=3.8.4
matplotlib-inline==0.1.7
meshio==4.4.6
mistune==3.1.2
mpmath==1.3.0
narwhals==1.30.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx~=3.2.1
notebook==7.3.2
notebook_shim==0.2.4
numpy~=1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
open3d==0.12.0
opt-einsum==3.3.0
overrides==7.7.0
packaging==24.2
pandas~=2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==6.0.0
plyfile==1.0.3
pooch==1.8.2
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==3.20.3
psutil~=5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
pyparsing==3.1.4
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-slugify==8.0.4
pytz==2025.1
pyvista==0.44.2
PyYAML~=6.0.1
pyzmq==26.3.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.1
rsa==4.9
scikit-learn~=1.5.0
scipy~=1.13.1
scooby==0.10.0
seaborn~=0.13.2
Send2Trash==1.8.3
six==1.17.0
sklearn==0.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.17.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.5.0
tinycss2==1.2.1
tomli==2.2.1
torch==2.4.1
tornado==6.4.2
tqdm~=4.66.4
traitlets==5.14.3
transforms3d==0.3.1
trimesh==3.9.14
triton==3.0.0
typeguard==2.13.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.8.0
tzdata==2025.1
uri-template==1.3.0
urllib3==2.2.3
utils==1.0.2
vtk==9.3.1
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.12.1
zipp==3.20.2

yaml~=0.2.5